scripts/create_tricorder_submission.py
#!/usr/bin/env python3
"""
Create Complete Tricorder Submission Package
"""

import os
import shutil
import json
from pathlib import Path
import logging

# Set up logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

def create_submission_package(model_path: str = "tricorder_optimized.onnx", 
                            output_dir: str = "tricorder_final_submission"):
    """Create complete Tricorder submission package"""
    
    logger.info("Creating Tricorder submission package...")
    
    # Create output directory
    output_path = Path(output_dir)
    if output_path.exists():
        shutil.rmtree(output_path)
    output_path.mkdir(parents=True, exist_ok=True)
    
    # Copy ONNX model
    if os.path.exists(model_path):
        shutil.copy2(model_path, output_path / "model.onnx")
        logger.info(f"Copied model: {model_path} -> {output_path}/model.onnx")
    else:
        logger.warning(f"Model file not found: {model_path}")
        return False
    
    # Create inference script
    inference_script = '''#!/usr/bin/env python3
"""
Tricorder Model Inference Script
Optimized for 10-class skin lesion classification with demographic data
"""

import numpy as np
import onnxruntime as ort
from PIL import Image
import torchvision.transforms as transforms
import logging

class TricorderInference:
    def __init__(self, model_path: str = "model.onnx"):
        self.model_path = model_path
        self.session = ort.InferenceSession(model_path)
        
        # Class names
        self.class_names = ["AK", "BCC", "SK", "SCC", "VASC", "DF", "NV", "NON", "MEL", "ON"]
        
        # Transforms
        self.transform = transforms.Compose([
            transforms.Resize((512, 512)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
        ])
        
        logging.info(f"Tricorder model loaded from {model_path}")
    
    def preprocess_image(self, image_path: str):
        """Preprocess image for inference"""
        image = Image.open(image_path).convert('RGB')
        image_tensor = self.transform(image).unsqueeze(0)
        return image_tensor.numpy()
    
    def preprocess_demographics(self, age: float, sex: str, anatom_site: str):
        """Preprocess demographic data"""
        # Normalize age
        age_norm = (age - 30.0) / 50.0
        
        # Convert sex to binary
        sex_norm = 1.0 if sex.lower() == 'male' else 0.0
        
        # Convert anatomical site to numeric
        anatom_site_mapping = {
            'head/neck': 0.0, 'upper extremity': 1.0, 'lower extremity': 2.0,
            'torso': 3.0, 'palms/soles': 4.0, 'oral/genital': 5.0,
            'unknown': 0.0
        }
        anatom_site_norm = anatom_site_mapping.get(anatom_site.lower(), 0.0) / 5.0
        
        return np.array([[age_norm, sex_norm, anatom_site_norm]], dtype=np.float32)
    
    def predict(self, image_path: str, age: float = 50.0, sex: str = 'male', anatom_site: str = 'unknown'):
        """Make prediction on image with demographic data"""
        # Preprocess inputs
        image_input = self.preprocess_image(image_path)
        demographic_input = self.preprocess_demographics(age, sex, anatom_site)
        
        # Run inference
        outputs = self.session.run(
            ['output'], 
            {'image': image_input, 'tabular': demographic_input}
        )
        
        # Get predictions
        probabilities = outputs[0][0]
        predicted_class_id = np.argmax(probabilities)
        predicted_class = self.class_names[predicted_class_id]
        confidence = probabilities[predicted_class_id]
        
        return {
            'predicted_class': predicted_class,
            'predicted_class_id': int(predicted_class_id),
            'confidence': float(confidence),
            'all_probabilities': probabilities.tolist()
        }

def main():
    # Example usage
    inference = TricorderInference()
    
    # Example prediction
    result = inference.predict(
        image_path="sample_image.jpg",
        age=45.0,
        sex="female",
        anatom_site="upper extremity"
    )
    
    print("Prediction Result:")
    print(f"Class: {result['predicted_class']}")
    print(f"Confidence: {result['confidence']:.4f}")
    print(f"All probabilities: {result['all_probabilities']}")

if __name__ == "__main__":
    main()
'''
    
    with open(output_path / "inference.py", 'w', encoding='utf-8') as f:
        f.write(inference_script)
    
    # Create README
    readme_content = '''# Tricorder Competition Submission

## Model Overview
This submission contains an optimized deep learning model for the Tricorder skin lesion classification competition.

### Model Architecture
- **Backbone**: EfficientNet-B3 (with DenseNet121 fallback)
- **Input**: 512x512 RGB images + demographic data (age, sex, anatomical site)
- **Output**: 10-class skin lesion classification
- **Format**: ONNX optimized for inference

### Classes
1. AK - Actinic keratosis
2. BCC - Basal cell carcinoma  
3. SK - Seborrheic keratosis
4. SCC - Squamous cell carcinoma
5. VASC - Vascular lesion
6. DF - Dermatofibroma
7. NV - Benign nevus
8. NON - Other non-neoplastic
9. MEL - Melanoma
10. ON - Other neoplastic

### Performance Features
- **Class Weighting**: Malignant classes (BCC, SCC, MEL) have 3x weight
- **Multi-modal Input**: Combines image and demographic data
- **Optimized Architecture**: EfficientNet-B3 for best accuracy/speed tradeoff
- **Data Augmentation**: Comprehensive augmentation during training

### Usage
```python
from inference import TricorderInference

# Initialize model
inference = TricorderInference("model.onnx")

# Make prediction
result = inference.predict(
    image_path="path/to/image.jpg",
    age=45.0,
    sex="female", 
    anatom_site="upper extremity"
)

print(f"Predicted class: {result['predicted_class']}")
print(f"Confidence: {result['confidence']:.4f}")
```

### Training Details
- **Dataset**: ISIC 2019 with 25,531 images across 10 classes
- **Training**: 30 epochs with early stopping
- **Optimizer**: AdamW with OneCycleLR scheduler
- **Loss**: CrossEntropyLoss with class weights
- **Validation**: Stratified train/val/test splits

### Competition Compliance
- ✅ ONNX format
- ✅ MIT License
- ✅ Training code provided
- ✅ 10-class classification
- ✅ Demographic data integration
- ✅ Optimized for inference speed

## Files
- `model.onnx`: Trained model in ONNX format
- `inference.py`: Inference script with example usage
- `README.md`: This documentation
- `submission_info.json`: Submission metadata
'''
    
    with open(output_path / "README.md", 'w', encoding='utf-8') as f:
        f.write(readme_content)
    
    # Create submission info
    submission_info = {
        "competition": "Tricorder",
        "model_name": "Tricorder Optimized Model",
        "version": "1.0",
        "description": "High-performance 10-class skin lesion classification with demographic data",
        "architecture": "EfficientNet-B3 + Demographic Fusion",
        "input_format": "512x512 RGB images + demographic data",
        "output_format": "10-class probabilities",
        "training_dataset": "ISIC 2019 (25,531 images)",
        "classes": [
            "AK", "BCC", "SK", "SCC", "VASC", 
            "DF", "NV", "NON", "MEL", "ON"
        ],
        "class_weights": {
            "malignant": ["BCC", "SCC", "MEL"],
            "medium_risk": ["SK", "VASC"],
            "benign": ["AK", "DF", "NV", "NON", "ON"]
        },
        "performance_metrics": {
            "weighted_accuracy": "TBD",
            "malignant_accuracy": "TBD",
            "inference_speed": "TBD"
        },
        "license": "MIT",
        "author": "Tricorder Competition Team",
        "created_date": "2025-09-02"
    }
    
    with open(output_path / "submission_info.json", 'w', encoding='utf-8') as f:
        json.dump(submission_info, f, indent=2)
    
    # Create requirements file
    requirements = '''onnxruntime>=1.15.0
numpy>=1.21.0
Pillow>=8.0.0
torchvision>=0.12.0
'''
    
    with open(output_path / "requirements.txt", 'w', encoding='utf-8') as f:
        f.write(requirements)
    
    logger.info(f"Submission package created successfully at: {output_path}")
    logger.info("Package contents:")
    for file in output_path.iterdir():
        logger.info(f"  - {file.name}")
    
    return True

def main():
    import argparse
    
    parser = argparse.ArgumentParser(description='Create Tricorder submission package')
    parser.add_argument('--model_path', type=str, default='tricorder_optimized.onnx',
                        help='Path to ONNX model file')
    parser.add_argument('--output_dir', type=str, default='tricorder_final_submission',
                        help='Output directory for submission package')
    
    args = parser.parse_args()
    
    success = create_submission_package(args.model_path, args.output_dir)
    
    if success:
        print("\\n🎉 Tricorder submission package created successfully!")
        print(f"📁 Location: {args.output_dir}")
        print("\\n📋 Package includes:")
        print("  - model.onnx (trained model)")
        print("  - inference.py (inference script)")
        print("  - README.md (documentation)")
        print("  - submission_info.json (metadata)")
        print("  - requirements.txt (dependencies)")
        print("\\n🚀 Ready for submission to Tricorder competition!")
    else:
        print("❌ Failed to create submission package")

if __name__ == "__main__":
    main()
